Fix crashing due to incorrect casts

+++ app/src/main/java/com/sprintkeyz/oxygenate/data/ConfigHelper.kt
@@ -42,7 +42,7 @@ inline fun <reified T> getPref(
         }
 
         percentage -> {
-            (value.toFloat() * 100f).toInt()
+            value.toFloat() * 100f
         }
 
         else -> value
